Allow Wordpress Auth to Return Arbitrary Data
Modify the WordpressAuth module so that instead of always returning the
User's ID, callers can specify what data to return upon successful
authentication - via the getUserData WordpressAuthConfig field.

For our application, returning the `Entity User` prevents repeating the
User fetching DB call in handlers. We may build & return a Map of the
UserMetas in the future(should see how expensive it is).