Support Custom Error Handling During WP Authentication
Add a WordpressAuthError type & a onAuthenticationFailure field to the
WordpressAuthConfig type. When auth validation fails, the field will be
called with the returned WordpressAuthError. This allows callers to use
a default type or return any http error(instead of being hardcoded to
our error status & messages).

Included with this is another AuthHandler, "wordpress-optional", which
uses the new failure field to return a `Maybe (Entity User)` instead of
throwing an HTTP error.