Add Support for Validation Wordpress REST Nonces
Modify the Servant AuthHandler to also check for valid nonces with the
"wp_rest" action. Nonces can be supplied via the `X-WP-Nonce` header or
`_wpnonce` query parameter.

The getUserData function in WordpressAuthConfig type now needs to return
the user id, since that is required for nonce validation.

Add new config fields for the nonce key & salt.

Add the validateWordpressNonce function to calculate & compare the nonce
to the expected hash substring.

Modify the wordpressHash & wordpressSalt to take an AuthScheme
specifying which key/salt pair to use.